PL-31: Set the site name based on the configuration values.
The form values had not been set yet, so I used the configuration values which is what the form is doing down the codebase.

I also set defaults in case it isn't found.

I'm not sure how to test this on the app yet and I reached out for more information.